fixbug:modify type of orderPayId Integer to Long

+++ ant-api/src/main/java/com/bx/ant/service/DriverOrderShopServiceI.java
@@ -80,7 +80,7 @@ public interface DriverOrderShopServiceI {
 	 * @param driverOrderPayId
 	 * @return
 	 */
-	DriverOrderShop getByPayId(Integer driverOrderPayId);
+	DriverOrderShop getByPayId(Long driverOrderPayId);
 
     /**
 	 * dataGrid.DriverOrderShopView
